Ransom note using hash tables
https://www.hackerrank.com/challenges/ctci-ransom-note/problem
Given the words present in a magazine and the words in a note, find whether the note can be made from the words in the magazine.

Create a hashmap for the words in the magazine. word,count
Similar hashmap for words in note.
For each word in note, find the count in the magazine hash. If magazine.count<note.count, return NO.