LuckBalance Greedy method
https://www.hackerrank.com/challenges/luck-balance/problem

There are few contests, their luck points and importance. When contest is lost, luck is added. When contest is won, luck is decreased. The aim is to maximize the luck. So lose as many contests as possible.
If the contest is not important, it can be lost. Also, there is a restriction of the maximum number of important contests that can be lost (given as k).